% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileManagement.R
\name{checkpack}
\alias{checkpack}
\title{Checks if a package is loaded}
\usage{
checkpack(pack)
}
\arguments{
\item{pack}{is a string indicating the name of the package to check.}
}
\value{
if the package is loaded then nothing happens. If not, then the package will be loaded.
}
\description{
Checks if a package is loaded
}
\examples{
checkpack('stats')
}

