% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateUtilities.R
\name{dates.create}
\alias{dates.create}
\title{Creates a sequence of dates}
\usage{
dates.create(startDate, endDate = Sys.Date(), labelDate = "Visit_Date",
  timeInterval = "1 day")
}
\arguments{
\item{startDate}{the begining date of the period}

\item{endDate}{the end date of the period}

\item{labelDate}{the label to be used for the date column}

\item{timeInterval}{the interval of the date sequence}
}
\value{
a data.frame with one date column
}
\description{
Creates a sequence of dates
}
\details{
The default interval is 1 day. The dates need to be well formatted.
  The default title for the date column is "Visit_Date". This function can be
  useful in creating timeseries.
}

