% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileManagement.R
\name{fetchFile}
\alias{fetchFile}
\title{Calls the file browser window for selecting a file}
\usage{
fetchFile()
}
\value{
: List of various information about the file selected by the user. The list contains the path and the name of the selected file.
It contains also some basic information about the file including its size and the date and time of file creation.
}
\description{
This function asks input from the user for selecting a file
}

