% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeModifiers.R
\name{permIntVec.1by1}
\alias{permIntVec.1by1}
\title{Reorders a vector of ordered integers.}
\usage{
permIntVec.1by1(alpha, omega, cnum)
}
\arguments{
\item{alpha}{An integer indicating the index of the item to be moved.}

\item{omega}{An integer indicating the index of the destination target integer.}

\item{cnum}{An integer indicating the number of items included in the vector of integers.}
}
\value{
A vector of the indexes of the columns in the new order.
}
\description{
Reorders a vector of ordered integers.
}

