% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeModifiers.R
\name{rename_cols}
\alias{rename_cols}
\title{Renames one or several columns in a data.frame}
\usage{
rename_cols(dat, oldnames, newnames)
}
\arguments{
\item{dat}{a data.frame containing the columns to be renamed.}

\item{oldnames}{a character vector of column names to be be renamed.}

\item{newnames}{a character vector of new names for the columns to be
renamed.}
}
\value{
the same data.frame with columns with new names.
}
\description{
Renames one or several columns in a data.frame
}
\details{
This function can be used in lieu of other functions in well know
  packages such as dplyr, plyr, and data.table. The squeleton of the function
  was found at:
  http://stackoverflow.com/questions/21502465/replacement-for-rename-in-dplyr.
}

