% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeModifiers.R
\name{savefile}
\alias{savefile}
\title{Saves to disc a dataframe as csv file}
\usage{
savefile(sfile, tfile, wdir = TRUE, rnames = FALSE)
}
\arguments{
\item{sfile}{The source file. An R dataframe, a matrix, or a vector}

\item{tfile}{The target csv file that will be saved to disc.}

\item{wdir}{A boolean variable indicating if the file will be saved in the working directory or in a user-selected directory will be used.}

\item{rnames}{A boolean variable indicating if row names will be saved in the csv file or not.}
}
\description{
Saves to disc a dataframe as csv file
}

