% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeModifiers.R
\name{xtable2df}
\alias{xtable2df}
\title{Converts a table to a data.frame}
\usage{
xtable2df(xtable, dname2ID = TRUE)
}
\arguments{
\item{xtable}{a table or an xtable object}

\item{dname2ID}{a boolean variable indicating if the row names should be used as the index for the new table.}
}
\value{
an R data frame with the data from the table plus the optional ID column made with the dimnames of the table
}
\description{
Converts a table to a data.frame
}
\details{
If dname2ID is true the dimnames will be converted into a column that can be considered as an index column.
If dname2ID is false then the dimnames of the tables remain as is. Once the dimnames have been converted into a column they are deleted.
}

